Add information section to projects page, so move it to shared components folder
