+++ src/app/projects/page.tsx
@@ -1,5 +1,6 @@
 import { Navbar } from '@/components/Shared/Navbar'
 import { Footer } from '@/components/Shared/Footer'
+import { Information } from '@/components/Shared/Information'
 import { Hero } from '@/components/Projects/Hero'
 import { ProjectsList } from '@/components/Projects/ProjectsList'
 
@@ -10,6 +11,7 @@ export default function Projects() {
       <main>
         <Hero />
         <ProjectsList />
+        <Information />
         <Footer />
       </main>
     </>
