+++ src/components/Shared/Information/index.tsx
